Fix errors where arguments should be double instead of int.
- We really should just get rid of set(int index, double value) in
geometry classes like Tuple2D

+++ IHMCGraphicsDescription/src/us/ihmc/graphicsDescription/yoGraphics/plotting/YoArtifactPosition.java
@@ -50,7 +50,7 @@ public class YoArtifactPosition extends YoArtifact
       graphics.setColor(color);
       graphics.setStroke(STROKE);
       
-      legendRadii.set(LEGEND_RADIUS, LEGEND_RADIUS);
+      legendRadii.set((double) LEGEND_RADIUS, (double) LEGEND_RADIUS);
       
       switch (graphicType)
       {
